GEOMESA-3400 Configurable Micrometer metrics - separate classloading
* Permits using metrics without unused dependencies on the classpath
