+++ geomesa-metrics/geomesa-metrics-micrometer/src/main/scala/org/locationtech/geomesa/metrics/micrometer/RegistryFactory.scala
@@ -1,0 +1,16 @@
+/***********************************************************************
+ * Copyright (c) 2013-2024 Commonwealth Computer Research, Inc.
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Apache License, Version 2.0
+ * which accompanies this distribution and is available at
+ * http://www.opensource.org/licenses/apache2.0.php.
+ ***********************************************************************/
+
+package org.locationtech.geomesa.metrics.micrometer
+
+import com.typesafe.config.Config
+import io.micrometer.core.instrument.MeterRegistry
+
+trait RegistryFactory {
+  def apply(conf: Config): MeterRegistry
+}
